Rename summarize to generate
